Merge PR resolves issues with authentication; authentication guards have been created to handle role based privileges.

+++ backend/src/recipes/recipes.service.ts
@@ -1,8 +1,8 @@
-import { Injectable } from '@nestjs/common';
+import { Injectable, InternalServerErrorException, NotFoundException } from '@nestjs/common';
 import { InjectRepository } from '@nestjs/typeorm';
 import { Recipes } from './recipes';
 import { Repository } from 'typeorm';
-import { CreateRecipeDto } from '../recipe-details/dto/create-recipe-dto';
+import { CreateRecipeDto } from '@/recipe-ingredients/dto/create-recipe-dto';
 
 @Injectable()
 export class RecipesService {
@@ -12,22 +12,34 @@ export class RecipesService {
   ) {}
 
   async allRecipes(): Promise<Recipes[]> {
-    return await this.recipesRepository.find();
+    return this.recipesRepository.find();
   }
 
   async findRecipe(recipeName: string): Promise<Recipes> {
-    return await this.recipesRepository.findOne({ where: { recipeName } });
+    return this.recipesRepository.findOne({ where: { recipeName } });
   }
 
   async addRecipe(recipeDto: CreateRecipeDto): Promise<Recipes> {
     const recipe = new Recipes();
     recipe.recipeName = recipeDto.recipeName;
     recipe.price = parseFloat(recipeDto.price);
-    return await this.recipesRepository.save(recipe);
+    recipe.mealType = recipeDto.mealType;
+    try { return this.recipesRepository.save(recipe);
+    } catch (error) {
+      throw new InternalServerErrorException(`Error adding recipe`);
+    }
   }
 
   async removeRecipe(recipeName: string): Promise<void> {
-    const recipe = await this.findRecipe(recipeName);
-    await this.recipesRepository.remove(recipe);
+    try {
+      const recipe = await this.findRecipe(recipeName);
+      if(recipe) {
+        await this.recipesRepository.remove(recipe);
+      } else {
+        throw new NotFoundException(`Recipe not found`);
+      }
+    } catch (error) {
+      throw new InternalServerErrorException(`Error removing recipe`);
+    }
   }
 }
